refactor(env): move kube details
